refactor: consolidate CryptoNetwork and KycInfo types in Onramp.ts
