--- src/types/CryptoNetwork.ts
@@ -1,13 +1,0 @@
-export enum CryptoNetwork {
-  bitcoin = 'bitcoin',
-  ethereum = 'ethereum',
-  solana = 'solana',
-  polygon = 'polygon',
-  stellar = 'stellar',
-  avalanche = 'avalanche',
-  base = 'base',
-  aptos = 'aptos',
-  optimism = 'optimism',
-  worldchain = 'worldchain',
-  xrpl = 'xrpl',
-}
